requirements.txt
terminaltables~=3.1.10
python-dotenv~=0.19.2
requests~=2.27.1